feat: add locations listings
